// projects/backoffice/src/app/private/drivers-operators/shared/interfaces/drivers.interfaces.ts
export class Drivers {
  id?: number;
  identityDocument: string;
  firstName: string;
  fatherSurname: string;
  motherSurname: string;
  genderId: number;
  birhtday: string;
  email: string;
  phone: string;
  address: string;
  currentLicences: any;
  documentsUrl: string;
  registrationState: boolean;
  document: any;
  areaId: number;
}